refactor: 예외 처리 패키지 구조 정리
